workspace_email_automation/requirements.txt
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-api-python-client==2.126.0
PyYAML==6.0.1
